# Optimize template replacement speed