--- a/Template.helpers.go
+++ b/Template.helpers.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"encoding/xml"
 	"log"
+	"strings"
 )
 
 // Convert given bytes to struct of xml nodes
@@ -118,4 +119,35 @@ func (t *Template) matchBrokenLeftPlaceholder(content string) bool {
 // Match right placeholder part `}}`
 func (t *Template) matchBrokenRightPlaceholder(content string) bool {
 	return t.matchBrokenPlaceholder(content, false)
+}
+
+func (t Template) GetContentPrefixList(content []byte) []string {
+	var ret []string
+	var record strings.Builder
+	start := false
+	length := len(content)
+	for i, v := range content {
+		if i == 0 {
+			continue
+		}
+
+		if v == '{' && content[i-1] == '{' {
+			start = true
+			continue
+		}
+		if start {
+			if v == ' ' || (v == '}' && length-1 > i && content[i+1] == '}') {
+				ret = append(ret, record.String())
+				record.Reset()
+				start = false
+			}
+			if v == '.' {
+				ret = append(ret, record.String())
+				record.Reset()
+				continue
+			}
+			record.WriteByte(v)
+		}
+	}
+	return ret
 }