1. fix ```clone()``` set parent method
2. optimize ```expandPlaceholders(xnode *xmlNode)``` ,reduce loop time
3. optimize ```replaceSingleParams(xnode *xmlNode, triggerParamOnly bool)```,reduce loop time

before:

```
goos: linux
goarch: amd64
pkg: github.com/bobiverse/docxplate
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
BenchmarkLists100-6     1000000000               0.4386 ns/op
BenchmarkLists200-6            1        1693713947 ns/op
BenchmarkLists400-6            1        6552133541 ns/op
BenchmarkLists1000-6           1        40482876671 ns/op
PASS
ok      github.com/bobiverse/docxplate  60.426s
```

after:

```
goos: linux
goarch: amd64
pkg: github.com/bobiverse/docxplate
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
BenchmarkLists100-6     1000000000               0.01451 ns/op
BenchmarkLists200-6     1000000000               0.02569 ns/op
BenchmarkLists400-6     1000000000               0.04924 ns/op
BenchmarkLists1000-6    1000000000               0.1187 ns/op
PASS
ok      github.com/bobiverse/docxplate  3.301s
```